move to shared lib
