+++ domains/certificates/Certificates.sln
@@ -19,6 +19,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "MeasurementEvents", "Shared
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Contracts", "Shared\Contracts\Contracts.csproj", "{03A0CD09-F291-4D30-9D8A-A2DE18142312}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AggregateRepositories", "Shared\AggregateRepositories\AggregateRepositories.csproj", "{0020310F-7891-4C33-9490-A48F8BC09981}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -49,6 +51,10 @@ Global
 		{03A0CD09-F291-4D30-9D8A-A2DE18142312}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{03A0CD09-F291-4D30-9D8A-A2DE18142312}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{03A0CD09-F291-4D30-9D8A-A2DE18142312}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0020310F-7891-4C33-9490-A48F8BC09981}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0020310F-7891-4C33-9490-A48F8BC09981}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0020310F-7891-4C33-9490-A48F8BC09981}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0020310F-7891-4C33-9490-A48F8BC09981}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -60,6 +66,7 @@ Global
 		{536410DD-6B50-47DC-99EF-CBCFBAAE0079} = {C500AE25-2567-4855-863B-7B51BC59F806}
 		{7CE965F2-DF46-4372-A44F-BA14957DC9C9} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
 		{03A0CD09-F291-4D30-9D8A-A2DE18142312} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
+		{0020310F-7891-4C33-9490-A48F8BC09981} = {50B52CB7-AA59-483B-8C89-46A681B3E01F}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {8710F924-9EE6-4142-A232-456121E239B0}
