+++ domains/certificates/Query.API/API/GranularCertificateIssuer/EnergyMeasuredConsumer.cs
@@ -1,6 +1,6 @@
 using System.Threading.Tasks;
+using AggregateRepositories;
 using API.ContractService;
-using API.GranularCertificateIssuer.Repositories;
 using CertificateEvents.Aggregates;
 using CertificateEvents.Primitives;
 using MassTransit;
