+++ domains/certificates/Shared/AggregateRepositories/AggregateRepositories.csproj
@@ -6,4 +6,12 @@
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Marten" Version="5.11.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\CertificateEvents\CertificateEvents.csproj" />
+  </ItemGroup>
+
 </Project>
