+++ domains/certificates/Shared/AggregateRepositories/AggregateRepositoryExtensions.cs
@@ -5,7 +5,7 @@ using System.Threading.Tasks;
 using CertificateEvents.Aggregates;
 using Marten;
 
-namespace API.GranularCertificateIssuer.Repositories;
+namespace AggregateRepositories;
 
 public static class AggregateRepositoryExtensions
 {
